+++ huey/tests/__init__.py
@@ -4,7 +4,6 @@ from huey.tests.test_crontab import *
 from huey.tests.test_django_tasks import *
 from huey.tests.test_helpers import *
 from huey.tests.test_immediate import *
-from huey.tests.test_kt_huey import *
 from huey.tests.test_mini import *
 from huey.tests.test_postgres import *
 from huey.tests.test_priority import *